Once again I did a faily large update. Sorry.
Fixed a small bug with the way GUI did a hittest, had forgot to null out active window upon failed testing, so you could still drag a window even thought the mouse was not over it.

Did a few small things for speed improvments..

The mian thing was added (hackishly, will be cleaned up) double and tripple click support for the mouse. On press a boolean is set (possibleClick) if you move the mouse while down this is set to false, this still allows for draging, but also allows for invalidation. Everytime a click happens a even is sent down the pipe (with a boolean telling the function if it is the final value) allowing the object to handel this if it is safe (or if it is the final ) the functions return true if the mouse should contiue handeling the clicks, otherwise false..

Also started a little bit of doxygen comments. More to come.

Signed-off-by: brandon <j.brandon.s@gmail.com>